セクションごとに実装を分けた
