+++ composeApp/src/commonMain/kotlin/org/khaki/schoolwatch/screens/clock/ClockScreen.kt
@@ -1,4 +1,4 @@
-package org.khaki.schoolwatch
+package org.khaki.schoolwatch.screens.clock
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -8,6 +8,11 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import org.khaki.schoolwatch.ClockTicker
+import org.khaki.schoolwatch.DigitalClockDisplay
+import org.khaki.schoolwatch.Schedule
+import org.khaki.schoolwatch.SushiSecondHandClock
+import org.khaki.schoolwatch.Task
 
 @Composable
 fun ClockScreen(
