+++ composeApp/src/commonMain/kotlin/org/khaki/schoolwatch/screens/settings/TaskManagementSection.kt
@@ -1,0 +1,122 @@
+package org.khaki.schoolwatch.screens.settings
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AddCircle
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import org.khaki.schoolwatch.Task
+import org.khaki.schoolwatch.localization.stringResource
+import org.khaki.schoolwatch.theme.DraculaColors
+
+@Composable
+fun TaskManagementSection(
+    tasks: List<Task>,
+    onAddTask: (String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    var textInput by remember { mutableStateOf("") }
+    
+    Column(modifier = modifier) {
+        Text(
+            text = stringResource().taskManagement,
+            style = MaterialTheme.typography.headlineMedium
+        )
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        // Task Management Section with background
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .background(
+                    color = DraculaColors.CurrentLine,
+                    shape = RoundedCornerShape(16.dp)
+                )
+                .padding(16.dp)
+        ) {
+            Column {
+                // Task addition section
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    OutlinedTextField(
+                        value = textInput,
+                        onValueChange = { textInput = it },
+                        modifier = Modifier.weight(1f),
+                        placeholder = { Text(stringResource().addTaskPlaceholder) },
+                        singleLine = true
+                    )
+                    Spacer(modifier = Modifier.width(8.dp))
+                    IconButton(
+                        onClick = {
+                            if (textInput.isNotBlank()) {
+                                onAddTask(textInput)
+                                textInput = ""
+                            }
+                        },
+                        enabled = textInput.isNotBlank()
+                    ) {
+                        Icon(
+                            Icons.Filled.AddCircle,
+                            contentDescription = stringResource().addTask,
+                            modifier = Modifier.size(36.dp),
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                    }
+                }
+
+                Spacer(modifier = Modifier.height(16.dp))
+
+                // Display current tasks
+                Text(
+                    text = stringResource().currentTasks,
+                    style = MaterialTheme.typography.titleLarge
+                )
+
+                Spacer(modifier = Modifier.height(8.dp))
+
+                if (tasks.isEmpty()) {
+                    Text(
+                        text = stringResource().noTasks,
+                        style = MaterialTheme.typography.bodyLarge
+                    )
+                } else {
+                    LazyColumn {
+                        items(
+                            items = tasks,
+                            key = { task -> task.id }
+                        ) { task ->
+                            Text(
+                                text = "• ${task.text}${if (task.isCompleted) " (${stringResource().taskCompleted})" else ""}",
+                                style = MaterialTheme.typography.bodyLarge,
+                                modifier = Modifier.padding(vertical = 4.dp)
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
